qa-6: Responsive design initial implementation, Folder restrcturing
