+++ lib/src/common/responsive_widget.dart
@@ -1,0 +1,51 @@
+import 'package:flutter/cupertino.dart';
+
+const int largeScreenSize = 1366;
+const int mediumScreenSize = 900;
+const int smallScreenSize = 768;
+const int customScreenSize = 1100;
+
+class ResponsiveWidget extends StatelessWidget {
+  final Widget largeScreen;
+  final Widget? mediumScreen;
+  final Widget? smallScreen;
+  const ResponsiveWidget({
+    Key? key,
+    required this.largeScreen,
+    this.mediumScreen,
+    this.smallScreen,
+  }) : super(key: key);
+
+  static bool isSmallestScreen(BuildContext context) =>
+      MediaQuery.of(context).size.width < smallScreenSize;
+
+  static bool isSmallScreen(BuildContext context) =>
+      MediaQuery.of(context).size.width < mediumScreenSize;
+
+  static bool isMediumScreen(BuildContext context) =>
+      MediaQuery.of(context).size.width >= mediumScreenSize &&
+      MediaQuery.of(context).size.width < largeScreenSize;
+
+  static bool isLargeScreen(BuildContext context) =>
+      MediaQuery.of(context).size.width >= largeScreenSize;
+
+  static bool isCustomScreen(BuildContext context) =>
+      MediaQuery.of(context).size.width > mediumScreenSize &&
+      MediaQuery.of(context).size.width <= customScreenSize;
+
+  @override
+  Widget build(BuildContext context) {
+    return LayoutBuilder(
+      builder: (context, constraints) {
+        final double _width = constraints.maxWidth;
+        if (_width >= largeScreenSize) {
+          return largeScreen;
+        } else if (_width >= mediumScreenSize) {
+          return mediumScreen ?? largeScreen;
+        } else {
+          return smallScreen ?? largeScreen;
+        }
+      },
+    );
+  }
+}
