+++ lib/src/features/questions/widgets/question_box.dart
@@ -1,0 +1,87 @@
+import 'dart:math';
+
+import 'package:flutter/material.dart';
+import 'package:quiz_app/src/api/models/models.dart';
+import 'package:quiz_app/src/features/questions/question_view.dart';
+
+class QuestionBox extends StatefulWidget {
+  const QuestionBox(
+      {Key? key, required this.question, required this.onQuestionAnswer})
+      : super(key: key);
+
+  final Question question;
+  final void Function(QuestionStatus, int) onQuestionAnswer;
+
+  @override
+  State<QuestionBox> createState() => _QuestionBoxState();
+}
+
+class _QuestionBoxState extends State<QuestionBox>
+    with AutomaticKeepAliveClientMixin {
+  int? _currentAnswer;
+  int? correctAnswerIndex;
+  bool isCorrect = false;
+
+  List<Widget> generateChoiceWidgets(Question question) {
+    correctAnswerIndex ??= correctAnswerIndex = Random().nextInt(3);
+    final widgetList = <ListTile>[];
+    for (var i in [for (var i = 0; i < 3; i += 1) i]) {
+      widgetList.add(ListTile(
+        contentPadding: EdgeInsets.zero,
+        leading: Radio(
+            value: i,
+            groupValue: _currentAnswer,
+            onChanged: (int? newValue) {
+              setState(() {
+                _currentAnswer = newValue!;
+
+                if (correctAnswerIndex == newValue) {
+                  isCorrect = true;
+                }
+              });
+              widget.onQuestionAnswer(
+                  correctAnswerIndex == newValue
+                      ? QuestionStatus.correct
+                      : QuestionStatus.wrong,
+                  question.id);
+            }),
+        title: Text(
+          i == correctAnswerIndex
+              ? question.correctAnswer
+              : question.incorrectAnswers[i],
+          style: const TextStyle(fontSize: 14),
+        ),
+      ));
+    }
+    return widgetList;
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    super.build(context);
+    return Padding(
+      padding: const EdgeInsets.all(8.0),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          Text(
+            widget.question.question,
+            style: Theme.of(context)
+                .textTheme
+                .headline6!
+                .copyWith(color: Colors.blue.shade900, fontSize: 18),
+          ),
+          const SizedBox(
+            height: 8,
+          ),
+          ...generateChoiceWidgets(widget.question)
+        ],
+      ),
+    );
+  }
+
+  @override
+  bool get wantKeepAlive => true;
+}
+// question screen
